% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.dna.r
\name{as.character}
\alias{as.character}
\alias{as.character.Dna}
\title{Coerce Dna to character string}
\usage{
\method{as.character}{Dna}(x, ...)
}
\arguments{
\item{x}{Dna object to be coerced to character}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Coerce Dna sequence to character string
}
\examples{
x <- Dna("acgtnacgtn")
as.character(x)

}
