% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.dna.r
\name{as.Dna.Dna}
\alias{as.Dna.Dna}
\title{Coerce Dna to Dna}
\usage{
\method{as.Dna}{Dna}(obj, code = obj$code, pattern = obj$pattern)
}
\arguments{
\item{obj}{object to coerce to Dna}

\item{code}{see \link{Dna}}

\item{pattern}{see \link{Dna}}
}
\description{
see \link{as.Dna}\cr
this is basically used to force Dna reencoding.
}
\examples{
x <- Dna("acgtacgt")
as.Dna(x, code="strict")

}
