% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.dna.r
\name{as.Dna}
\alias{as.Dna}
\title{Generic method to coerce to Dna}
\usage{
as.Dna(obj, code, pattern)
}
\arguments{
\item{obj}{object to coerce to Dna}

\item{code}{see \link{Dna}}

\item{pattern}{see \link{Dna}}
}
\value{
Dna object
}
\description{
coerce character string or Dna object to Dna object\cr
this is basically the same as the \link{Dna} constructor.\cr
for Dna object this allows to force a re-encoding
}
\seealso{
\link{Dna}, \link{as.Dna.Dna}, \link{as.Dna.character}
}
