% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.dna.r
\name{as.ri.Dna}
\alias{as.ri.Dna}
\title{Coerce Dna to ri range [1, length(dna)]}
\usage{
\method{as.ri}{Dna}(obj)
}
\arguments{
\item{obj}{object to coerce to \link[bit]{ri}}
}
\description{
Coerce Dna to ri range [1, length(dna)]
}
\seealso{
\link{as.ri}, \link[bit]{ri}
}
