% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.dna.r
\name{as.ri}
\alias{as.ri}
\title{Generic method to coerce to range}
\usage{
as.ri(obj)
}
\arguments{
\item{obj}{object to coerce to \link[bit]{ri}}
}
\description{
coerce to range-index (\link{ri}) object from library \link{bit}
}
\seealso{
\link[bit]{ri} in library \link{bit}, \link{as.ri.Dna}
}
