% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.dna.r
\name{as.ri.ri}
\alias{as.ri.ri}
\title{Coerce bit::ri to bit::ri}
\usage{
\method{as.ri}{ri}(obj)
}
\arguments{
\item{obj}{object to coerce to \link[bit]{ri}}
}
\description{
just a trivial helper (absent from bit)
}
