% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.bbaf.r
\name{baf.count.filter}
\alias{baf.count.filter}
\title{filter allele counts}
\usage{
baf.count.filter(count, lowread = 2L, mincov = 10L, minall = 1L,
  maxall = 2L, deltafreq = 0.1, what = c("count", "index",
  "logical"))
}
\arguments{
\item{count}{nx4 integer matrix of alleles counts.
as returned by \code{baf.fetch.xxx}, \code{baf.sample.xxx} or
\code{baf.heterozygous.xxx}.}

\item{lowread}{low read threshold (see details)}

\item{mincov}{minimal coverage (see details)}

\item{minall}{minimal number of alleles (see details)}

\item{maxall}{maximal number of alleles (see details)}

\item{deltafreq}{max allele frequency (see details)}

\item{what}{kind of result to return (see value)}
}
\value{
\itemize{
\item if (what == "count") filtered count matrix
\item if (what == "index") indexes of valid rows
\item if (what == "logical") logical vector indicating valid rows
}
}
\description{
filter allele counts according to coverage, min and max
number of alleles and max allele frequency (see details).
}
\details{
\bold{definitions}\cr
an allele X is present iff:\cr
\itemize{\item count_X > \code{lowread}}
a site (i.e. a row of \code{count}) is valid iff:\cr
\itemize{
\item coverage >= \code{covmin}
\item \code{minall} <= nb_alleles <= \code{maxall}
\item if (nb_alleles > 1) abs(0.5 - count_max_allele / coverage) <= \code{deltafreq}
}

if \code{mincov} is < 0 then \code{mincov} is computed as:\cr
\code{quantile(coverage, -mincov/1000)}\cr

if \code{delta.freq} == NA or maxall < 2 then freq condition is ignored
}
\examples{
baf <- baf.open(lx.system.file("samples/test.baf", "xlx"))
cnt <- baf.sample(baf, sample.size=Inf)
cnt.ok <- baf.count.filter(cnt, lowread=2, mincov=10, minall=1, maxall=4, deltafreq=NA)
baf.close(baf)
}
