% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.bbaf.r
\name{baf.count.genotype}
\alias{baf.count.genotype}
\title{get genotype from allele counts}
\usage{
baf.count.genotype(count, lowread = 2L, what = c("symbol", "index",
  "string"), sorted = FALSE, use.threads = lx.use.threads())
}
\arguments{
\item{count}{nx4 integer matrix of alleles counts.
as returned by \code{baf.fetch.xxx}, \code{baf.sample.xxx} or
\code{baf.heterozygous.xxx}.}

\item{lowread}{low read threshold (see details)}

\item{what}{kind of result to return (see value)}

\item{sorted}{sort result by frequencies (see details)}

\item{use.threads}{(see \link{lx.use.threads})}
}
\value{
list of alleles or array of strings (see details)
}
\description{
retrieve genotypes from allele counts matrix
}
\details{
this function returns the list of allele(s) present at each
row of the \code{count} matrix.\cr
an allele X is present iff: \code{count_X > lowread}\cr
the typeof result depends upon the \code{what} parameter:\cr
\itemize{
\item if (what == "symbol") list of character array of alleles symbols
\item if (what == "index") list of integer array of alleles indices
\item if (what == "string") array of genotype strings
}
if \code{sorted}==FALSE (default) each array element of the result list
(or each character in string) appears in the same order as in \code{colnames(count)}
whatever the frequency value.
if \code{sorted}==TRUE then array elements are sorted by decreasing frequency.
}
\examples{
baf <- baf.open(lx.system.file("samples/test.baf", "xlx"))
cnt <- baf.sample(baf, sample.size=Inf)
# sample: count all genotypes
cnt.ok <- baf.count.filter(cnt, lowread=0, mincov=10, minall=1, maxall=4, deltafreq=NA)
geno <- baf.count.genotype(cnt.ok, lowread=0)
table(nb.all <- sapply(geno, length))
# sample: ordered bi-allelic genotypes
cnt.ok <- baf.count.filter(cnt, lowread=0, mincov=10, minall=2, maxall=2, deltafreq=NA)
geno <- baf.count.genotype(cnt.ok, lowread=0, sorted=TRUE)
baf.close(baf)
}
