% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.bbaf.r
\name{baf.count.regularize.gaussian}
\alias{baf.count.regularize.gaussian}
\title{allelic frequency regularization}
\usage{
baf.count.regularize.gaussian(count, tot, sd = 0.5, n = 10L,
  .seed = -1L)
}
\arguments{
\item{count}{nx4 integer matrix of alleles counts.
as returned by \code{baf.fetch.xxx}, \code{baf.sample.xxx} or}

\item{tot}{total count (i.e. depth)}

\item{sd}{gaussian standard deviation}

\item{n}{number of draws per point}

\item{.seed}{seed for random (do not seed if < 0)}
}
\value{
names list of 3 vectors (cnt=regularized counts,
tot=regularized total, raf=regulized all. freq.)
}
\description{
regularize allelic frequency using gaussian pseudo-counts
}
\examples{
N <- 10000
cnt <- round(rpois(N, 5))
tot <- cnt + round(rpois(N, 5))
unreg <- cnt / tot
reg <- baf.count.regularize(cnt, tot, "gaussian", .seed=0)

}
