% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.bbaf.r
\name{baf.count.regularize.poisson}
\alias{baf.count.regularize.poisson}
\title{allelic frequency regularization}
\usage{
baf.count.regularize.poisson(count, tot, n = 10L, .eps = 0,
  .seed = -1L)
}
\arguments{
\item{count}{nx4 integer matrix of alleles counts.
as returned by \code{baf.fetch.xxx}, \code{baf.sample.xxx} or
\code{baf.heterozygous.xxx}.}

\item{tot}{total count (i.e. depth)}

\item{n}{number of draws per point}

\item{.eps}{lambda correction}

\item{.seed}{seed for random (do not seed if < 0)}
}
\value{
names list of 3 vectors (cnt=regularized counts,
tot=regularized total, raf=regulized all. freq.)
}
\description{
regularize allelic frequency using poisson pseudo-counts
}
\examples{
N <- 10000
cnt <- round(rpois(N, 5))
tot <- cnt + round(rpois(N, 5))
unreg <- cnt / tot
reg <- baf.count.regularize(cnt, tot, "poisson", .seed=0)

}
