% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.bbaf.r
\name{baf.count.regularize}
\alias{baf.count.regularize}
\title{allelic frequency regularization}
\usage{
baf.count.regularize(count, tot, method = c("poisson", "gaussian"), ...)
}
\arguments{
\item{count}{nx4 integer matrix of alleles counts.
as returned by \code{baf.fetch.xxx}, \code{baf.sample.xxx} or}

\item{tot}{total count (i.e. depth)}

\item{method}{method to use ("poisson" or "gaussian")}

\item{...}{specific method parameters}
}
\value{
names list of 3 vectors (cnt=regularized counts,
tot=regularized total, raf=regularized all. freq.)
}
\description{
regularize allelic frequency using various methods
}
\examples{
N <- 10000
cnt <- round(rpois(N, 30))
tot <- cnt + round(rpois(N, 30))
unreg <- cnt / tot
reg <- baf.count.regularize(cnt, tot, "gaussian", .seed=0)

}
