% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.bbaf.r
\name{baf.heterozygous.chr}
\alias{baf.heterozygous.chr}
\title{get heterozygous positions on chromosome}
\usage{
baf.heterozygous.chr(handle, chr, lowread = 2L, mincov = 10L,
  deltafreq = 0.1, .chunk.size = NA, .sample.size = 1000000L)
}
\arguments{
\item{handle}{file handle (as returned by \link{baf.open})}

\item{chr}{chromosome index (if integer) or chromosome name (if character)}

\item{lowread}{low read threshold (see \link{baf.count.filter})}

\item{mincov}{minimal coverage (see \link{baf.count.filter})}

\item{deltafreq}{max allele frequency (see \link{baf.count.filter})}

\item{.chunk.size}{chunk size for loading chromosome (to save memory). Use NA
to load in one single chunk (quicker but memory expensive). see note below for the
compatibility with a negative mincov.}

\item{.sample.size}{sample size to determine mincov for the case where .chunk.size != NA
and mincov < 0 (see note).}
}
\value{
integer matrix of size n x 4 containing allele counts at heterozygous sites.
with (1-based) positions as rownames.
}
\description{
get heterozygous positions on chromosome \code{chr}.
this function gather all heterozygous positions defined 
as valid by \link{baf.count.filter}:\cr
\itemize{
\item coverage >= \code{covmin}
\item nb_alleles == 2
\item abs(0.5 - count_max_allele / coverage) <= \code{deltafreq}
}
}
\note{
if (mincov < 0) then mincov will be estimated 
(\code{quantile(coverage, -mincov/1000)}) as in \link{baf.count.filter}.
However the region for computing coverage will depends upon the \code{.chunk.size}
parameter: if \code{.chunk.size == NA} then this will be performed on the
whole chromosome.
if \code{.chunk.size != NA} then mincov will be first evaluated
on a sample of \code{.sample.size} data points. Therefore for exact results, you
should better use a positive value for mincov.

see \link{HELP.COORD} for help on coordinates systems
}
\examples{
baf <- baf.open(lx.system.file("samples/test.baf", "xlx"))
baf.heterozygous.chr(baf, 3, deltafreq=0.5)
baf.close(baf)

}
\seealso{
\link{baf.heterozygous.coord}, \link{baf.heterozygous.cloc}
}
