% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.bbaf.r
\name{baf.heterozygous.coord}
\alias{baf.heterozygous.coord}
\title{get heterozygous positions within region}
\usage{
baf.heterozygous.coord(handle, coord, lowread = 2L, mincov = 10L,
  deltafreq = 0.1)
}
\arguments{
\item{handle}{file handle (as returned by \link{baf.open})}

\item{coord}{absolute sequence coordinates (c(absfrom, absto)) (1-based)
or a single position absfrom (this implies absto=absfrom)}

\item{lowread}{low read threshold (see \link{baf.count.filter})}

\item{mincov}{minimal coverage (see \link{baf.count.filter})}

\item{deltafreq}{max allele frequency (see \link{baf.count.filter})}
}
\value{
integer matrix of size n x 4 containing allele counts at heterozygous sites.
with (1-based) positions as rownames.
}
\description{
get heterozygous positions within region using absolute coordinates.\cr
\code{coord} defines a region on a chromosome. this function
gather all heterozygous positions defined as valid by \link{baf.count.filter}:\cr
\itemize{
\item coverage >= \code{covmin}
\item nb_alleles == 2
\item abs(0.5 - count_max_allele / coverage) <= \code{deltafreq}
}
}
\note{
take care that if mincov < 0 the actual mincov will be computed on this
region (as \code{quantile(coverage, -mincov/1000)}),
not on whole chromosome nor genome. you better use a positive value for
mincov.

see \link{HELP.COORD} for help on coordinates systems
}
\examples{
baf <- baf.open(lx.system.file("samples/test.baf", "xlx"))
baf.heterozygous.coord(baf, c(33725, 33732), deltafreq=0.5)
baf.close(baf)

}
\seealso{
\link{baf.heterozygous.cloc}, \link{baf.count.filter}
}
