% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.bbaf.r
\name{baf.index2name}
\alias{baf.index2name}
\title{convert seqindex to seqname}
\usage{
baf.index2name(handle, seqindex, silent = FALSE)
}
\arguments{
\item{handle}{basta/baf file handle (as returned by \link{basta.open}
or \link{baf.open})}

\item{seqindex}{integer vector of 1-based sequence index}

\item{silent}{logical keep silent if index(es) out of bounds}
}
\description{
convert seqindex to seqname.\cr
see \link{basta.index2name} this is the same function
}
