% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.bbaf.r
\name{baf.name2index}
\alias{baf.name2index}
\title{convert seqname to seqindex}
\usage{
baf.name2index(handle, seqname, silent = FALSE)
}
\arguments{
\item{handle}{basta/baf file handle (as returned by \link{basta.open}
or \link{baf.open})}

\item{seqname}{character vector of sequence name(s)}

\item{silent}{logical keep silent if seqname(s) not found}
}
\description{
convert seqname to seqindex.\cr
see \link{basta.name2index} this is the same function
}
