% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.bbaf.r
\name{baf.open}
\alias{baf.open}
\title{open baf file}
\usage{
baf.open(filename)
}
\arguments{
\item{filename}{baf file name}
}
\value{
baf file handle
}
\description{
open baf file for reading
}
\examples{
baf <- baf.open(lx.system.file("samples/test.baf", "xlx"))
baf.close(baf)

}
