% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.bbaf.r
\name{baf.sample}
\alias{baf.sample}
\title{get sample of alleles counts}
\usage{
baf.sample(handle, chrs = NULL, sample.size = 1000000L, .seed = -1L,
  use.threads = lx.use.threads())
}
\arguments{
\item{handle}{file handle (as returned by \link{baf.open})}

\item{chrs}{integer (chromosome indexes) or character (chromosome names) vector
specifying which chromosomes to use. Use NULL to specify all
chromosome declared in baf header.}

\item{sample.size}{sample size (set to +Inf to collect all positions)}

\item{.seed}{random seed (for reproducibility). use a strictly negative
integer to disable seeding.}

\item{use.threads}{(see \link{lx.use.threads})}
}
\value{
integer matrix of size sample.size x 4 containing allele counts.
with rownames of the form 'chrindex.position'.
}
\description{
sample positions on all chromosomes and return allele counts
}
\note{
you may set sample.size to +Inf to collect allele counts on all position.
but be careful this may use a very large amount of memory.
}
\examples{
baf <- baf.open(lx.system.file("samples/test.baf", "xlx"))
smp <- baf.sample(baf, sample.size=100, .seed=0, use.threads=FALSE)
baf.close(baf)

}
