% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.bbaf.r
\name{baf2clocs}
\alias{baf2clocs}
\title{make clocations spanning all chromosomes}
\usage{
baf2clocs(handle)
}
\arguments{
\item{handle}{basta/baf file handle (as returned by \link{basta.open}
or \link{baf.open})}
}
\value{
nx3 matrix of clocations
}
\description{
make clocations spanning all chromosomes declared in basta/baf file
}
\details{
this is the same function as \link{basta2clocs}
}
\note{
see \link{HELP.COORD} for help on coordinates systems
}
