% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.bbaf.r
\name{baf2coords}
\alias{baf2coords}
\title{make absolute coordinates spanning all chromosomes}
\usage{
baf2coords(handle)
}
\arguments{
\item{handle}{basta/baf file handle (as returned by \link{basta.open}
or \link{baf.open})}
}
\value{
nx2 matrix of absolute coordinates (1-based)
}
\description{
make absolute coordinates spanning all chromosomes
declared in basta/baf file
}
\details{
this is the same function as \link{basta2coords}
}
\note{
see \link{HELP.COORD} for help on coordinates systems
}
