% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.basta.r
\name{basta.count.cloc}
\alias{basta.count.cloc}
\title{fetch symbols counts using relative clocation}
\usage{
basta.count.cloc(handle, clocation, truncate = TRUE, sym = c("A", "C",
  "G", "T", "other"), binsize = clocation[3] - clocation[2] + 1,
  case.sensitive = FALSE, drop = TRUE)
}
\arguments{
\item{handle}{basta file handle (as returned by \link{basta.open})}

\item{clocation}{relative clocation = c(seqname, from, to) (1-based)}

\item{truncate}{truncate 3' to seq.size if needed}

\item{sym}{vector of strings specifying symbols to be counted.
see details.}

\item{binsize}{size of bins (defaults to whole sequence)}

\item{case.sensitive}{symbols in \code{sym} are case sensitive}

\item{drop}{drop the last window if sequence length is not a muliple
of binsize}
}
\value{
a matrix or vector of counts. if \code{length(sym)==1} returns
a vector of symbol(s) counts for each bin position.
if \code{length(sym)>1} returns a matrix of symbols counts
with length(sym) columns and each row corresponds to each bin
position.
}
\description{
fetch sequence thru \link{basta.fetch.cloc} and
count the number of occurences of symbols specified
in \code{sym} in contiguous bins of size \code{binsize}
}
\details{
each string in \code{sym} specifies a set of symbols to be
counted. if this set starts with '!', it means symbols \bold{not} in
set. As a special case the string "Other" is equivalent to "!ACGT".
}
\note{
see \link{HELP.COORD} for help on coordinates systems
}
\examples{
fh <- basta.open(lx.system.file("samples/test.bst", "xlx"))
basta.fetch.cloc(fh, c(1, 1, 24))
# count all DNA symbols
basta.count.cloc(fh, c(1, 1, 24))
# count GC only
basta.count.cloc(fh, c(1, 1, 24), sym="GC")
# count GC in bins
basta.count.cloc(fh, c(1, 1, 24), sym="GC", binsize=4)
basta.close(fh)
}
\seealso{
\link{basta.count.coord}
}
