% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.basta.r
\name{basta.count.coord}
\alias{basta.count.coord}
\title{fetch symbols counts using absolute coordinates}
\usage{
basta.count.coord(handle, coord, sym = c("A", "C", "G", "T", "other"),
  binsize = diff(range(coord)) + 1, case.sensitive = FALSE,
  drop = TRUE)
}
\arguments{
\item{handle}{basta file handle (as returned by \link{basta.open})}

\item{coord}{absolute coordinates (c(absfrom, absto)) (1-based) or
single absolute position.}

\item{sym}{vector of strings specifying symbols to be counted.
see details.}

\item{binsize}{size of bins (defaults to whole sequence length)}

\item{case.sensitive}{symbols in \code{sym} are case sensitive}

\item{drop}{drop the last bin if sequence length is not a muliple
of binsize}
}
\value{
a matrix or vector of counts. if \code{length(sym)==1} returns
a vector of symbol(s) counts for each bin position.
if \code{length(sym)>1} returns a matrix of symbols counts
with length(sym) columns and each row corresponds to each bin
position.
}
\description{
fetch sequence thru \link{basta.fetch.coord} and
count the number of occurences of symbols specified
in \code{sym} in contiguous bins of size \code{binsize}
}
\details{
each string in \code{sym} specifies a set of symbols to be
counted. if this set starts with '!', it means symbols \bold{not} in
set. As a special case the string "Other" is equivalent to "!ACGT".
}
\note{
see \link{HELP.COORD} for help on coordinates systems
}
\examples{
fh <- basta.open(lx.system.file("samples/test.bst", "xlx"))
basta.fetch.coord(fh, c(1, 24))
# count all DNA symbols
basta.count.coord(fh, c(1, 24))
# count GC only
basta.count.coord(fh, c(1, 24), sym="GC")
# count GC in bins
basta.count.coord(fh, c(1, 24), sym="GC", binsize=4)
basta.close(fh)
}
\seealso{
\link{basta.count.coord}
}
