% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.basta.r
\name{basta.fetch.cloc}
\alias{basta.fetch.cloc}
\title{fetch sequence using relative clocation}
\usage{
basta.fetch.cloc(handle, clocation, truncate = TRUE)
}
\arguments{
\item{handle}{basta file handle (as returned by \link{basta.open})}

\item{clocation}{relative clocation = c(seqname, from, to) (1-based)}

\item{truncate}{truncate 3' to seq.size if needed}
}
\value{
sequence string
}
\description{
fetch sequence using relative clocation
}
\note{
see \link{HELP.COORD} for help on coordinates systems
}
\examples{
fh <- basta.open(lx.system.file("samples/test.bst", "xlx"))
basta.fetch.coord(fh, c(25, 34))
basta.fetch.sloc(fh, "seq2:1-10")
basta.fetch.cloc(fh, c(2, 1, 10))
basta.close(fh)

}
