% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.basta.r
\name{basta.fetch.coord}
\alias{basta.fetch.coord}
\title{fetch sequence using absolute coordinates}
\usage{
basta.fetch.coord(handle, coord)
}
\arguments{
\item{handle}{basta file handle (as returned by \link{basta.open})}

\item{coord}{absolute coordinates (c(absfrom, absto)) (1-based) or
single absolute position.}
}
\value{
sequence string
}
\description{
fetch sequence using absolute coordinates
}
\note{
see \link{HELP.COORD} for help on coordinates systems
}
\examples{
fh <- basta.open(lx.system.file("samples/test.bst", "xlx"))
basta.fetch.coord(fh, c(25, 34))
basta.fetch.sloc(fh, "seq2:1-10")
basta.fetch.cloc(fh, c(2, 1, 10))
basta.close(fh)

}
