% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.basta.r
\name{basta.fetch.sloc}
\alias{basta.fetch.sloc}
\title{fetch sequence using relative slocation}
\usage{
basta.fetch.sloc(handle, slocation, zero.based.loc = FALSE,
  truncate = TRUE)
}
\arguments{
\item{handle}{basta file handle (as returned by \link{basta.open})}

\item{slocation}{relative slocation ("seqname:from-to")}

\item{zero.based.loc}{given slocation is 0-based}

\item{truncate}{truncate 3' to seq.size if needed}
}
\value{
sequence string
}
\description{
fetch sequence using relative slocation
}
\note{
see \link{HELP.COORD} for help on coordinates systems
}
\examples{
fh <- basta.open(lx.system.file("samples/test.bst", "xlx"))
basta.fetch.coord(fh, c(25, 34))
basta.fetch.sloc(fh, "seq2:1-10")
basta.fetch.cloc(fh, c(2, 1, 10))
basta.close(fh)

}
