% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.basta.r
\name{basta.index2name}
\alias{basta.index2name}
\title{convert seqindex to seqname}
\usage{
basta.index2name(handle, seqindex, silent = FALSE)
}
\arguments{
\item{handle}{basta/baf file handle (as returned by \link{basta.open}
or \link{baf.open})}

\item{seqindex}{integer vector of 1-based sequence index}

\item{silent}{logical keep silent if index(es) out of bounds}
}
\value{
character vector of seq name(s) or NULL if index(es) out of bounds
}
\description{
convert seqindex to seqname
}
\examples{
fh <- basta.open(lx.system.file("samples/test.bst", "xlx"))
basta.name2index(fh, "seq1")
basta.index2name(fh, 2)
basta.name2index(fh, c("seq1", "nothere"))
basta.index2name(fh, 1:3)
basta.close(fh)

}
\seealso{
\link{basta.name2index}
}
