% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.basta.r
\name{basta.open}
\alias{basta.open}
\title{open basta file}
\usage{
basta.open(filename, check.crc32 = FALSE)
}
\arguments{
\item{filename}{basta file name}

\item{check.crc32}{perform crc32 check}
}
\value{
basta file handle
}
\description{
open basta file for reading
}
\examples{
fh <- basta.open(lx.system.file("samples/test.bst", "xlx"), check.crc32=TRUE)
basta.close(fh)

}
