% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.basta.r
\name{basta2clocs}
\alias{basta2clocs}
\title{make clocations spanning all chromosomes}
\usage{
basta2clocs(handle)
}
\arguments{
\item{handle}{basta/baf file handle (as returned by \link{basta.open}
or \link{baf.open})}
}
\value{
nx3 matrix of clocations
}
\description{
make clocations spanning all chromosomes declared in basta/baf file
}
\note{
see \link{HELP.COORD} for help on coordinates systems
}
\examples{
fh <- basta.open(lx.system.file("samples/test.bst", "xlx"))
clocs <- basta2clocs(fh)
basta.close(fh)

}
