% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.basta.r
\name{basta2coords}
\alias{basta2coords}
\title{make absolute coordinates spanning all chromosomes}
\usage{
basta2coords(handle)
}
\arguments{
\item{handle}{basta/baf file handle (as returned by \link{basta.open}
or \link{baf.open})}
}
\value{
nx2 matrix of absolute coordinates (1-based)
}
\description{
make absolute coordinates spanning all chromosomes
declared in basta/baf file
}
\note{
see \link{HELP.COORD} for help on coordinates systems
}
\examples{
fh <- basta.open(lx.system.file("samples/test.bst", "xlx"))
coords <- basta2coords(fh)
basta.close(fh)

}
