% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.bed.r
\name{bed2clocs}
\alias{bed2clocs}
\title{convert bed regions to matrix of clocations}
\usage{
bed2clocs(handle, bed, check = TRUE)
}
\arguments{
\item{handle}{basta/baf file handle (as returned by \link{basta.open}
or \link{baf.open}}

\item{bed}{dataframe (as returned by \link{bed.read})}

\item{check}{check that boundaries are correct}
}
\value{
a nx3 matrix of (1-based) clocations
}
\description{
see \link{HELP.COORD} for help on coordinates systems
}
\examples{
bed <- bed.read(lx.system.file("samples/test.bed", "xlx"))
fh <- basta.open(lx.system.file("samples/test.bst", "xlx"))
clocs <- bed2clocs(fh, bed)
basta.close(fh)

}
