% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.dna.r
\name{c}
\alias{c}
\alias{c.Dna}
\title{Catenate two or more Dna sequences}
\usage{
\method{c}{Dna}(obj, ...)
}
\arguments{
\item{obj}{Dna object}

\item{...}{Dna objects or character strings (may be mixed)}
}
\value{
a Dna sequence
}
\description{
Catenate two or more Dna sequences
}
\examples{
x <- Dna("acgtnacgtn")
c(x, "rryy", Dna("gg"))

}
