% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.coords.r
\name{cloc2coord}
\alias{cloc2coord}
\title{transform relative clocation to absolute coordinates}
\usage{
cloc2coord(handle, clocation, truncate = TRUE)
}
\arguments{
\item{handle}{basta/baf file handle (as returned by \link{basta.open}
or \link{baf.open})}

\item{clocation}{relative clocation c(chrindex, from, to) (1-based)}

\item{truncate}{truncate 3' to seq.size if needed}
}
\value{
absolute coordinates c(absfrom, absto) (1-based), NULL on error
}
\description{
see \link{HELP.COORD} for help on coordinates systems
}
\examples{
fh <- basta.open(lx.system.file("samples/test.bst", "xlx"))
cloc2coord(fh, c(1, 1, 10))
cloc2coord(fh, c(2, 1, 10))
basta.close(fh)

}
