% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.coords.r
\name{cloc2sloc}
\alias{cloc2sloc}
\title{transform relative clocation to relative slocation}
\usage{
cloc2sloc(handle, clocation, zero.based.loc = FALSE)
}
\arguments{
\item{handle}{basta/baf file handle (as returned by \link{basta.open}
or \link{baf.open})}

\item{clocation}{relative clocation c(chrindex, from, to) (1-based)}

\item{zero.based.loc}{returned slocation should be 0-based}
}
\value{
relative slocation "chrname:from-to" (0 or 1-based), NULL on error
}
\description{
see \link{HELP.COORD} for help on coordinates systems
}
\examples{
fh <- basta.open(lx.system.file("samples/test.bst", "xlx"))
cloc2sloc(fh, c(1, 1, 10))
basta.close(fh)

}
