% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.coords.r
\name{clocations}
\alias{clocations}
\title{create a matrix of clocations from data}
\usage{
clocations(x = NULL)
}
\arguments{
\item{x}{data to reformat (see details)}
}
\value{
nx3 matrix of clocations with proper
colnames and storage.
}
\description{
check if data is a proper matrix of clocations and
reformat it if necessary.
}
\details{
data can be :\cr
\itemize{
\item NULL : return empty matrix
\item matrix : (should be nx3) then just setup colnames and storage mode
\item dataframe : (should be nx3) then convert to matrix 
\item  anything else: transform to nx3 matrix 
}
}
\examples{
clocations() # empty clocs
clocations(list(c(1,1,10), c(2,1,10)))
clocations(c(1,1,10, 2,1,10))
fh <- basta.open(lx.system.file("samples/test.bst", "xlx"))
clocations(lapply(c("seq1:1-10", "seq2:1-10"), sloc2cloc, handle=fh))
basta.close(fh)
}
\seealso{
\link{clocs.matrix}
}
