% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.coords.r
\name{clocs.included}
\alias{clocs.included}
\title{assess inclusion from one set of clocations to another}
\usage{
clocs.included(clocations, include, use.threads = lx.use.threads())
}
\arguments{
\item{clocations}{nx3 matrix of clocations to be tested.}

\item{include}{nx3 matrix of clocations to search in}

\item{use.threads}{(see \link{lx.use.threads})}
}
\value{
logical vector of length nrow(clocations)
}
\description{
given two sets of clocations, indicates for each clocation of
the first set if it is included in at least one clocation of the
second set.
}
\examples{
c1 <- clocations(c(1,1,10))
c2 <- clocations(c(1,1,5, 1,6,10))
clocs.included(c1, c2)
clocs.included(c2, c1)
c1 <- clocations(c(1,1,10, 1,20,30, 2,10,20))
c2  <- clocations(c(1,1,20, 1,25,30))
clocs.included(c1, c2)
clocs.included(c2, c1)
clocs.included(c1, NULL)
clocs.included(NULL, NULL)
# the difference with clocs.covered is in the reduction of the second argument
c1 <- clocations(c(1,20,30))
c2 <- clocations(c(1,1,25, 1,25,40))
clocs.included(c1, c2)
clocs.covered(c1, c2)
clocs.included(c1, clocs.reduce(c2))

}
\seealso{
\link{clocs.covered}, \link{clocs.overlap}
}
