% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.coords.r
\name{clocs.is.disjoint}
\alias{clocs.is.disjoint}
\title{test if clocations are disjoint}
\usage{
clocs.is.disjoint(clocations, strong = TRUE)
}
\arguments{
\item{clocations}{nx3 matrix of relative clocations (1-based)}

\item{strong}{if strongly disjoint}
}
\value{
boolean TRUE if (weakly/strongly) disjoint
}
\description{
Test if clocations are (weakly or strongly) disjoint.\cr
weakly disjoint <=> no interval is completely
included in another one (but intervals may overlap)\cr
strongly disjoint <=> no two intervals overlap.
}
\note{
this function runs slightly quicker if clocations has already
been sorted by \link{clocs.sort} with decreasing.to=TRUE.
}
\examples{
clocs <- clocations(c(1,1,10, 1,11,20, 1,21,30, 2,5,10))
clocs.is.disjoint(clocs)
clocs <- clocations(c(1,1,10, 1,5,20, 1,10,30, 2,5,10))
clocs.is.disjoint(clocs)
clocs.is.disjoint(clocs, strong=FALSE)
clocs <- clocations(c(1,1,10, 1,5,30, 1,10,30, 2,5,10))
clocs.is.disjoint(clocs, strong=FALSE)

}
