% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.coords.r
\name{clocs.is.empty}
\alias{clocs.is.empty}
\title{test if clocations is empty}
\usage{
clocs.is.empty(clocations)
}
\arguments{
\item{clocations}{nx3 matrix of relative clocations (1-based)}
}
\value{
boolean TRUE if empty.
}
\description{
Test if clocations is empty (either null or no rows)
}
\examples{
clocs.is.empty(NULL)
clocs.is.empty(clocations())
clocs.is.empty(clocations(c(1,1,10)))

}
