% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.coords.r
\name{clocs.is.unsorted}
\alias{clocs.is.unsorted}
\title{test if clocations are \bold{not} sorted}
\usage{
clocs.is.unsorted(clocations, decreasing.to = FALSE)
}
\arguments{
\item{clocations}{nx3 matrix of relative clocations (1-based)}

\item{decreasing.to}{boolean or NA.
should the sort order of \code{to} be increasing or decreasing?
may be set to NA if you don't care about the order of to.}
}
\value{
boolean TRUE if unsorted, FALSE if sorted
}
\description{
Test if clocations are not sorted without the cost of sorting it.
}
\details{
the sort order is: first by increasing chromosome index
(clocations[,1]) then, for equal chromosome index, by increasing from
(clocations[,2]) then, for equal from, by increasing to if 
\code{decreasing.to==FALSE} else by decreasing to.
}
\note{
for coords you may use the R base function \link{is.unsorted}
}
\examples{
clocs <- clocations(c(1,1,5, 1,10,10, 1,10,20, 2,5,10))
clocs.is.unsorted(clocs)
clocs.is.unsorted(clocs, decreasing.to=TRUE)
clocs.is.unsorted(clocs, decreasing.to=NA)
clocs <- clocs.sort(clocs, decreasing.to=TRUE)

}
\seealso{
\link{clocs.sort}
}
