% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.coords.r
\name{clocs.is.valid}
\alias{clocs.is.valid}
\title{clocations sanity check}
\usage{
clocs.is.valid(clocations, handle = NULL)
}
\arguments{
\item{clocations}{nx3 matrix of relative clocations (1-based)}

\item{handle}{basta/baf file handle (as returned by \link{basta.open}
or \link{baf.open}). this parameter is optional (see details).}
}
\value{
boolean TRUE if valid.
}
\description{
Test if clocations matrix is valid i.e. that\cr
\code{1 <= from <= to <= seq.len} and \code{1 <= chr <= nchr}.
}
\details{
if \code{handle} is provided then the function checks
that \code{to <= seq.len} and \code{chr <= nchr} else these
conditions are ignored.
}
\note{
an empty clocations is valid.
}
\examples{
fh <- basta.open(lx.system.file("samples/test.bst", "xlx"))
clocs.is.valid(NULL)
clocs <- clocations(c(1,10,25, 2,30,10))
clocs.is.valid(clocs)
clocs <- clocations(c(1,10,25, 2,10,30))
clocs.is.valid(clocs)
clocs.is.valid(clocs, fh)
clocs <- clocations(c(1,10,24, 2,10,30))
clocs.is.valid(clocs, fh)
basta.close(fh)

}
