% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.coords.r
\name{clocs.matrix}
\alias{clocs.matrix}
\title{reformat data to proper matrix of clocations}
\usage{
clocs.matrix(x = NULL)
}
\arguments{
\item{x}{data to reformat (see details)}
}
\value{
nx3 matrix of clocations with proper
colnames and storage.
}
\description{
this function is mostly used within other functions to properly
(re)format a clocs matrix. It is quite unusual to call it directly,
please consider \link{clocations} instead.
}
\details{
data can be :\cr
\itemize{
\item NULL : return empty matrix
\item matrix : (should be nx3) then just setup colnames and storage mode
\item dataframe : (should be nx3) then convert to matrix 
\item  anything else: transform to nx3 matrix 
}
}
\examples{
clocs.matrix() # empty clocs
clocs.matrix(list(c(1,1,10), c(2,1,10)))
clocs.matrix(c(1,1,10, 2,1,10))
fh <- basta.open(lx.system.file("samples/test.bst", "xlx"))
clocs.matrix(lapply(c("seq1:1-10", "seq2:1-10"), sloc2cloc, handle=fh))
basta.close(fh)
}
\seealso{
\link{clocations}
}
