% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.coords.r
\name{clocs.overlap}
\alias{clocs.overlap}
\title{get overlaps from one set of clocations to another}
\usage{
clocs.overlap(clocations, over, use.threads = lx.use.threads())
}
\arguments{
\item{clocations}{nx3 matrix of relative clocations (1-based)}

\item{over}{mx3 matrix of relative clocations (1-based)}

\item{use.threads}{(see \link{lx.use.threads})}
}
\value{
a list of length nrow(clocations). the elements are vector
of indices indicating which elements of the second set are overlapped.
A list element of length 0 indicates no overlapping elements.
}
\description{
given two sets of clocations, indicates for each clocation of
the first set which clocations from the second set it overlaps
}
\note{
multithreading is performed by chromosomes
}
\examples{
c1 <- clocations(c(1,1,10, 1,20,30, 2,1,10))
clocs.overlap(c1, c1)
c2 <- clocations(c(1,1,20))
clocs.overlap(c1, c2)
clocs.overlap(c2, c1)
c2 <- clocations(c(1,1,15, 1,5,25))
clocs.overlap(c1, c2)
clocs.overlap(c2, c1)
c2 <- clocations(c(2,1,10, 1,1,15, 1,5,25))
clocs.overlap(c1, c2)
clocs.overlap(c2, c1)
clocs.overlap(c1, NULL)
clocs.overlap(NULL, NULL)

}
\seealso{
\link{clocs.nearest}, \link{clocs.included}, \link{clocs.covered}
}
