% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.coords.r
\name{clocs.rbind}
\alias{clocs.rbind}
\title{catenate a list of matrices of clocations into single matrix}
\usage{
clocs.rbind(submatrices)
}
\arguments{
\item{submatrices}{list of submatrices of clocations}
}
\value{
matrix of clocations
}
\description{
catenate (by rows) a list of matrices of clocations into a
single matrix of clocations and optionaly reorder rows\cr
}
\details{
this is equivalent to \code{do.call(rbind, submatrices)}
followed by an optional reordering of rows when the \code{submatrices}
argument has a \code{split} attribute (see \link{clocs.rsplit}).
}
\examples{
fh <- basta.open(lx.system.file("samples/test.bst", "xlx"))
clocs <- coords2clocs(fh, 1:25)
x <- clocs.rsplit(clocs)
y <- clocs.rbind(x)
z <- clocs2coords(fh, y)
identical(as.integer(z[,1]), 1:25)
clocs[1,1] <- 2L
clocs[25,1] <- 1L
x <- clocs.rsplit(clocs, keep=T)
y <- clocs.rbind(x)
identical(y, clocs)
basta.close(fh)

}
\seealso{
\link{clocs.rsplit} for the reverse operation
}
