% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.coords.r
\name{clocs.rsplit}
\alias{clocs.rsplit}
\title{split matrix of clocations into submatrices}
\usage{
clocs.rsplit(clocations, by = clocations[, 1], keep.split = FALSE)
}
\arguments{
\item{clocations}{matrix of clocations}

\item{by}{group factors (should be of length nrow(clocations))}

\item{keep.split}{logical keep split indexes as attribute (see details)}
}
\value{
list of submatrices (named by factors)
}
\description{
split matrix of clocations (by rows) into submatrices according
to \code{by} \cr
}
\details{
if \code{keep.split} is TRUE then the original indexes of clocations
are kept (as attribute \code{split}) for further reordering (see \link{clocs.rbind}).
}
\note{
levels in \code{by} that do not occur are dropped.
}
\examples{
fh <- basta.open(lx.system.file("samples/test.bst", "xlx"))
clocs <- coords2clocs(fh, 1:25)
x <- clocs.rsplit(clocs)
y <- clocs.rbind(x)
z <- clocs2coords(fh, y)
identical(as.integer(z[,1]), 1:25)
basta.close(fh)

}
\seealso{
\link{clocs.rbind} for the reverse operation
}
