% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.coords.r
\name{clocs.sort}
\alias{clocs.sort}
\title{sort clocations}
\usage{
clocs.sort(clocations, decreasing.to = FALSE, order = FALSE)
}
\arguments{
\item{clocations}{nx3 matrix of relative clocations (1-based)}

\item{decreasing.to}{boolean.
should the sort order of \code{to} be decreasing or increasing?}

\item{order}{return order index instead of clocations}
}
\value{
sorted nx3 matrix of relative clocations (if order=FALSE)
or an integer vector of row indexes (if order=TRUE)
}
\description{
sort clocations in increasing order of chr, from and increasing or
decreasing order of to.
}
\details{
the sort order is: first by increasing chromosome index
(clocations[,1]) then, for equal chromosome index, by increasing from
(clocations[,2]) then, for equal from, by increasing to
if \code{decreasing.to=FALSE} else by decreasing to.
}
\note{
for coords you may simply use the R base function \link{sort}
}
\examples{
clocs <- clocations(c(1,10,20, 1,10,30, 2,5,10, 1,1,100))
clocs.sort(clocs)
clocs.sort(clocs, decreasing.to=TRUE)

}
\seealso{
\link{clocs.is.unsorted}
}
