% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.coords.r
\name{clocs.threshold}
\alias{clocs.threshold}
\title{filter clocations by size}
\usage{
clocs.threshold(clocations, minsize = 1L, reduce = TRUE,
  use.threads = lx.use.threads())
}
\arguments{
\item{clocations}{nx3 matrix of relative clocations (1-based)}

\item{minsize}{minimum width}

\item{reduce}{perform a \link{clocs.reduce} first}

\item{use.threads}{(see \link{lx.use.threads}) (only used if reduce==TRUE)}
}
\description{
keep only clocations which size \code{(w=to-from+1)} is greater
or equal to \code{minsize}.
}
\examples{
c1 <- clocations(c(1,1,5, 1,20,30, 2,1,10))
identical(clocs.threshold(c1, minsize=0), c1)
clocs.threshold(c1, minsize=10)
clocs.threshold(c1, minsize=30)
clocs.threshold(NULL)

}
