% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.coords.r
\name{clocs2coords}
\alias{clocs2coords}
\title{transform matrix of relative clocations to matrix of absolute coordinates}
\usage{
clocs2coords(handle, clocations)
}
\arguments{
\item{handle}{basta/baf file handle (as returned by \link{basta.open}
or \link{baf.open})}

\item{clocations}{nx3 matrix of relative clocation (1-based)}
}
\value{
nx2 matrix of absolute coordinates  (1-based)
}
\description{
transform a nx3 matrix of relative clocations to
a nx2 matrix of absolute coordinates\cr
see \link{HELP.COORD} for help on coordinates systems
}
\examples{
fh <- basta.open(lx.system.file("samples/test.bst", "xlx"))

x <- sample(1:50, 10, replace=TRUE)
coo <- cbind(x, x)
clocs <- coords2clocs(fh, coo)
plocs <- coords2clocs(fh, x)
identical(clocs, plocs)

x <- sample(25:30, 10, replace=TRUE)
coo <- cbind(from=x, to=x+10)
clocs <- coords2clocs(fh, coo)
rcoo <-clocs2coords(fh, clocs)
identical(coo, rcoo)

clocs <- coords2clocs(fh, matrix(0, ncol=2, nrow=0))
clocs2coords(fh, clocs)
clocs <- coords2clocs(fh, matrix(1, ncol=2, nrow=1))
clocs2coords(fh, clocs)

basta.close(fh)

}
