% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.coords.r
\name{clocs2llocs}
\alias{clocs2llocs}
\title{convert clocations to llocations}
\usage{
clocs2llocs(clocations)
}
\arguments{
\item{clocations}{nx3 matrix of relative clocations (1-based)}
}
\value{
named list of mx2 relative locations per chromosome
}
\description{
convert a nx3 matrix of clocations to a list of
mx2 locations.
(see \link{HELP.COORD} for help on coordinates systems)
}
\examples{
fh <- basta.open(lx.system.file("samples/test.bst", "xlx"))
clocs <- coords2clocs(fh, 1:25)
llocs <- clocs2llocs(clocs)
rclocs <- llocs2clocs(llocs)
identical(clocs, rclocs)
basta.close(fh)

}
