% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.dna.r
\name{compl.Dna}
\alias{compl.Dna}
\title{Complement Dna sequence}
\usage{
\method{compl}{Dna}(obj)
}
\arguments{
\item{obj}{a Dna sequence to complement}
}
\description{
just complement sequence (not reverse complement)
}
\examples{
x <- Dna("acgtnry")
compl(x)

}
\seealso{
\link{revcompl.Dna}
}
