% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.dna.r
\name{compl}
\alias{compl}
\title{Generic method to complement a sequence}
\usage{
compl(obj)
}
\arguments{
\item{obj}{a Dna sequence to complement}
}
\description{
just complement sequence (not reverse complement)
}
\seealso{
\link{compl.Dna}, \link{revcompl}
}
