% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.coords.r
\name{coord2cloc}
\alias{coord2cloc}
\title{transform absolute coordinates to relative clocation}
\usage{
coord2cloc(handle, coord, truncate = TRUE)
}
\arguments{
\item{handle}{basta/baf file handle (as returned by \link{basta.open}
or \link{baf.open})}

\item{coord}{absolute coordinates c(absfrom, absto) (1-based) or
single absolute position}

\item{truncate}{truncate 5' and 3' to seq.size if needed}
}
\value{
relative clocation c(chrindex, from, to) (1-based), NULL on error
}
\description{
see \link{HELP.COORD} for help on coordinates systems
}
\examples{
fh <- basta.open(lx.system.file("samples/test.bst", "xlx"))
coord2cloc(fh, c(1, 10))
coord2cloc(fh, c(25, 34))
basta.close(fh)

}
