% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.coords.r
\name{coord2sloc}
\alias{coord2sloc}
\title{transform absolute coordinates to relative slocation}
\usage{
coord2sloc(handle, coord, zero.based.loc = FALSE, truncate = TRUE)
}
\arguments{
\item{handle}{basta/baf file handle (as returned by \link{basta.open}
or \link{baf.open})}

\item{coord}{absolute coordinates c(absfrom, absto) (1-based) or
single absolute position}

\item{zero.based.loc}{given slocation is 0-based}

\item{truncate}{truncate 3' to seq.size if needed}
}
\value{
relative slocation "chrname:from-to" (0 or 1-based), NULL on error
}
\description{
see \link{HELP.COORD} for help on coordinates systems
}
\examples{
fh <- basta.open(lx.system.file("samples/test.bst", "xlx"))
coord2sloc(fh, c(1, 10))
coord2sloc(fh, c(25, 34))
basta.close(fh)

}
