% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.coords.r
\name{coords.sample}
\alias{coords.sample}
\title{sample absolute point locations on chromosomes}
\usage{
coords.sample(handle, clocations, size = 1000000L, replace = FALSE)
}
\arguments{
\item{handle}{basta/baf file handle (as returned by \link{basta.open}
or \link{baf.open})}

\item{clocations}{regions where we can sample (endpoints included)}

\item{size}{number of points to sample}

\item{replace}{sample with replacement (see \link{lx.sample})}
}
\value{
vector of \code{size} absolute point coordinates.
(see \link{coords2clocs} to transform into clocations)
}
\description{
sample locations within regions specified by clocations.
}
\note{
if replace=FALSE and N, the number of points in the union
of all regions, is less than 2*size then downsample to N/2 points
}
\examples{
fh <- basta.open(lx.system.file("samples/test.bst", "xlx"))
clocs <- basta2clocs(fh)
samp <- coords.sample(fh, clocs, size=10)
coords2clocs(fh, sort(samp))
basta.close(fh)
}
