% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.coords.r
\name{coords2clocs}
\alias{coords2clocs}
\title{transform matrix of absolute coordinates to matrix of relative clocations}
\usage{
coords2clocs(handle, coords, truncate = TRUE)
}
\arguments{
\item{handle}{basta/baf file handle (as returned by \link{basta.open}
or \link{baf.open})}

\item{coords}{nx2 matrix of absolute coordinates  (1-based) or
vector of n absolute point coordinates}

\item{truncate}{truncate 5' and 3' to seq.size if needed}
}
\value{
nx3 matrix of relative clocation (1-based)
}
\description{
transform a nx2 matrix of absolute coordinates (or a vector of point
coordinates) to nx3 matrix of relative clocations.\cr
see \link{HELP.COORD} for help on coordinates systems
}
\note{
if some absolute coordinates span several chromosomes
then the corresponding rows are discarded.
}
\examples{
fh <- basta.open(lx.system.file("samples/test.bst", "xlx"))

x <- sample(1:50, 10, replace=TRUE)
coo <- cbind(x, x)
clocs <- coords2clocs(fh, coo)
plocs <- coords2clocs(fh, x)
identical(clocs, plocs)

x <- sample(25:30, 10, replace=TRUE)
coo <- cbind(from=x, to=x+10)
clocs <- coords2clocs(fh, coo)
rcoo <-clocs2coords(fh, clocs)
identical(coo, rcoo)

clocs <- coords2clocs(fh, matrix(0, ncol=2, nrow=0))
clocs <- coords2clocs(fh, matrix(1, ncol=2, nrow=1))

basta.close(fh)
}
