% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.dna.r
\name{countsymb.Dna}
\alias{countsymb.Dna}
\title{Count symbols in Dna Sequence}
\usage{
\method{countsymb}{Dna}(obj, symb = "acgt")
}
\arguments{
\item{obj}{a \link{Dna} sequence}

\item{symb}{a character string containing symbols to count}
}
\description{
Count symbols in Dna Sequence
}
\note{
To count using IUPAC degenerated codes, use uppercase
symbols (eg symb="W" will count g + c).
Therefore symb="r" will count the number of strict r's
whereas "R" will sum counts for r's and g's and a's.
}
\examples{
seq <- paste(sample(c("a", "c", "g", "t"), 1e7, replace=TRUE), collapse="")
system.time(cnt <- length(gregexpr("a", seq)[[1]]))
x <- Dna(seq, "strict")
system.time(cnt <- countsymb(x, "a"))
system.time(tab <- countsymb(x))
countsymb(x, "W")

}
