% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.dna.r
\name{countsymb}
\alias{countsymb}
\title{Generic method to count symbols in sequence}
\usage{
countsymb(obj, symb)
}
\arguments{
\item{obj}{sequence object (usually \link{Dna})}

\item{symb}{a character string containing symbols to count}
}
\value{
count table
}
\description{
Generic method to count symbols in sequence
}
\seealso{
\link{countsymb.Dna} for Dna sequence
}
