% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.dna.r
\name{Dna}
\alias{Dna}
\title{Dna Class constructor}
\usage{
Dna(x = 0L, code = c("standard", "strict", "pattern"),
  pattern = NULL)
}
\arguments{
\item{x}{either an integer, a character string or another Dna object}

\item{code}{encoding scheme (see details)}

\item{pattern}{pattern to use (only used if code="pattern")}
}
\description{
Dna Class constructor
}
\details{
\subsection{x parameter}{
if x is an integer, creates an empty Dna sequence of size x\cr
if x is a character string, creates a Dna sequence representing x\cr
if x is a Dna object, same as \code{Dna(as.character(x), ...)}\cr
}
\subsection{code parameter}{
if code == "standard" use 3 bits/symbol to represent
symbols in [acgtryn]\cr
if code == "strict" use a 2 bits/symbol to represent
symbols in [acgt]\cr
if code == "pattern" use a 2 bits/symbol to represent
pattern by [x.]\cr
}
}
\examples{
x <- Dna("acgtnnactg")
x <- Dna("acgtacgt", "strict")
x <- Dna("acgtnnnacgtg", "pattern", "[gc]")
x <- Dna(10)
x[] <- "accn"
x <- Dna(10, "strict")
x <- Dna(10, "pattern", "[gc]")
x[] <- "gca"

}
