% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.dna.r
\name{hamming.Dna}
\alias{hamming.Dna}
\title{Hamming distance between two Dna sequences}
\usage{
\method{hamming}{Dna}(obj1, obj2)
}
\arguments{
\item{obj1}{a \link{Dna} sequence}

\item{obj2}{a \link{Dna} sequence of same size and encoding as obj1}
}
\value{
number of differences between obj1 and obj2 or -1 if obj1 and obj2
are not of same size or encoding
}
\description{
Hamming distance between two Dna sequences
}
\note{
sequence case and IUPAC codes are ignored
}
\examples{
hamming(Dna("acgtacgtacgt"), Dna("acggacggacgg"))

}
