% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.dna.r
\name{hamming}
\alias{hamming}
\title{Generic method to compute Hamming distance between two sequences}
\usage{
hamming(obj1, obj2)
}
\arguments{
\item{obj1}{first sequence object (usually \link{Dna})}

\item{obj2}{second sequence object (usually \link{Dna})}
}
\description{
Generic method to compute Hamming distance between two sequences
}
\seealso{
\link{hamming.Dna} for Dna sequence
}
