% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.bbaf.r
\docType{data}
\name{HELP.BAF}
\alias{HELP.BAF}
\title{Baf format}
\description{
Baf is a binary format to compactly represents reads alignments
from bam files. It basically only keeps the information about
each allele counts at each position of the chromosomes
(all other information such as alignement quality, CIGAR etc...
is discarded). It can therefore be used to retrieve allelic frequencies,
total cover (sum of all alleles) or GC content (sum of G and C alleles).

This is a little endian binary file composed of
a header:\cr
\tabular{lll}{
int32 \tab 0x62696d31 \tab magic number ('baf1')\cr
int32 \tab nbseq \tab number of sequences\cr
---- \tab ------------ \tab -------\cr
\tab \tab repeat nbseq times\cr
---- \tab ------------ \tab -------\cr
string \tab namei \tab name of sequence i (see 1)\cr
int64 \tab sizei \tab length of sequence i\cr
int32 \tab codei \tab encoding size in bytes for this sequence (0, 1,2,4)\cr
---- \tab ------------ \tab -------\cr
}
followed by the concatenation of nbseq arrays each of
4 * sizei * codei bytes. each codei bytes (codei=1,2 or4)
represent total number of read bases (i.e non counting deletions) covering 
each position.\cr
(codei=1: unsigned char, codei=2: unsigned short, codei=4: unsigned int32,
and codei=0 means that all counts on this chromosome are 0)\cr.

(1) string format is:
\tabular{lll}{
int32 \tab size \tab string length\cr
bytes \tab size + 1 \tab NULL terminated char array\cr
}
}
\note{
When opening a baf file, the header is loaded
into memory but not the count arrays. Counts
will be directly accessed from disk when needed.

Conversion from bam to baf is performed by the
external C executable \code{bam2baf} provided
in Csrc directory.

Baf header is compatible with Basta header (see \link{HELP.BASTA})
and Baf handles can therefore be passed as the handle argument of
most functions accepting a Basta handle (except of course for
those that need to access to sequence).

of course when using together a Basta and a Baf file, you should ensure
that sequences in both header are stricly identical. In practice this means that
the Bam file from which the baf file was generated was built using
the same fasta sequences from which the Basta file was generated.\cr
}
