% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.basta.r
\docType{data}
\name{HELP.BASTA}
\alias{HELP.BASTA}
\title{Basta format}
\description{
Basta file format is similar to Fasta format
but allow indexed access to sequences.\cr
this is a little endian binary file composed of
a header:\cr
\tabular{lll}{
int32 \tab 0x62617332 \tab magic number ('bas2')\cr
int32 \tab nbseq \tab number of sequences\cr
---- \tab ------------ \tab -------\cr
\tab \tab repeat nbseq times\cr
---- \tab ------------ \tab -------\cr
string \tab namei \tab name of sequence i (see 1)\cr
int64 \tab sizei \tab length of sequence i\cr
int32 \tab crc32i \tab crc32 of sequence i\cr
---- \tab ------------ \tab -------\cr
}
followed by the concatenation of all sequences
as character arrays (not NULL terminated).

(1) string format is:
\tabular{lll}{
int32 \tab size \tab string length\cr
bytes \tab size + 1 \tab NULL terminated char array\cr
}
}
\note{
When opening a basta file, the header is loaded
into memory but not the sequences. Sequences
are directly accessed from disk.

Conversion from fasta to basta is performed by the
external C executable \code{fasta2basta} provided
in Csrc directory.
}
