% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.dna.r
\docType{data}
\name{HELP.DNA.ENCODING}
\alias{HELP.DNA.ENCODING}
\title{Dna Internal Encoding Scheme}
\description{
this section describes the internal Dna bits encoding schemes
and is intended for developpers only.\cr

\tabular{ll}{
\bold{code}\tab\bold{length(bits)}\cr
standard\tab 3\cr
strict\tab 2\cr
pattern\tab 1\cr
}

\subsection{standard encoding}{
\tabular{ll}{
\bold{i}\tab\bold{bits[[i]]}\cr
1\tab[acgt]\cr
2\tab[cgny]\cr
3\tab[gtnr]\cr
}
\tabular{llll}{
\bold{symb}\tab\bold{bit1}\tab\bold{bit2}\tab\bold{bit3}\cr
g\tab 1\tab 1\tab 1\cr
c\tab 1\tab 1\tab 0\cr
t\tab 1\tab 0\tab 1\cr
a\tab 1\tab 0\tab 0\cr
\tab\tab\tab\cr
n\tab 0\tab 1\tab 1\cr
y\tab 0\tab 1\tab 0\cr
r\tab 0\tab 0\tab 1\cr
x\tab 0\tab 0\tab 0\cr
}}

\subsection{strict encoding}{
\tabular{ll}{
\bold{i}\tab\bold{bits[[i]]}\cr
1\tab[gc]\cr
2\tab[gt]\cr
}
\tabular{lll}{
\bold{symb}\tab\bold{bit1}\tab\bold{bit2}\cr
g\tab 1\tab 1\cr
c\tab 1\tab 0\cr
t\tab 0\tab 1\cr
a\tab 0\tab 0\cr
}}

\subsection{pattern encoding}{
\tabular{ll}{
\bold{i}\tab\bold{bits[[i]]}\cr
1\tab pattern\cr
}
\tabular{ll}{
\bold{symb}\tab\bold{bit1}\cr
pattern\tab 1\cr
!pattern\tab 0\cr
}}
}
