% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.dna.r
\docType{data}
\name{HELP.DNA}
\alias{HELP.DNA}
\title{Dna Class}
\description{
\bold{Dna} is an S3 Class that let you manipulate DNA sequences
with a more memory-efficient way than usual character strings.\cr
More precisely \bold{Dna} stores sequence with a 3, 2 or 1 bits/symbol
instead of 8 bits for character strings.\cr

\bold{Dna} currently works with the following restrictions:
\itemize{
  \item the DNA alphabet is lowercase and restricted to "acgtryn" or "acgt"
        depending upon the storage mode.
  \item maximum sequence size is 2^31-1 (this is the same limitation as for
        character strings, until R internally goes to 64 bits vector indexes)
}
\subsection{creation}{
Dna sequences are created by the \link{Dna} constructor or \link{as.Dna} coercion.
}
\subsection{manipulation}{
Dna sequences can be transformed to strings by \link{as.character}\cr
Access to sequence components is performed by subscripting (either as
extracting or replacing):\cr
\code{dna[index]} and \code{dna[index] <- seq}\cr
}
\subsection{misc}{
other \bold{Dna} operations include:\cr
\link{c}, \link{length}, \link{subseq}, \link{summary},
\link{rev}, \link{compl}, \link{revcompl}, \link{plot} etc.
}
}
\note{
\bold{Dna} requires the \link{bit} library

the \code{print} S3 method has been redefined. For debugging purpose, you may
use \code{unclass(obj)} to see the actual internal components.
}
\examples{
# generate a 10 Mb sequence
n <- 1e7
seq <- paste(sample(c("a", "c", "g", "t"), n, replace=TRUE), collapse="")
x <- Dna(seq, "strict")
length(x)
summary(x)
s <- as.character(x)  # identical to seq
# extract or replace subsequences 
x[1:20]
x[20:1]
x[1:20] <- "acgt"
x[1:20] <- "acgtn"  # will complain
x[]  # same as Dna(x) or as.Dna(x) or x
# some operations
revcompl(x)
countsymb(x, "gc")
countsymb(x, "W")
plot(x) # guess what 
# some funny constructors
x <- Dna(15)
x[] <- "acg"
as.character(x[seq.int(1, length(x), by=3)])

}
