% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdb.obo.r
\docType{data}
\name{HELP.OBO}
\alias{HELP.OBO}
\title{lx in-memory database parser for Obo format}
\description{
read and parse Obo database in flat file format and hold
results in memory

the main functions are : \code{\link{mdb.obo.read}} and \code{\link{mdb.obo.load}}
}
\note{
these functions hold all the database in memory and are therefore not intented for
large databases
}
\examples{

db <- mdb.obo.load(lx.system.file("samples/test_obo", "xlx"))

# get entry names

names(db)

# get info about specific entry :

db$GO.0000001
db$GO.0000001$id
db$GO.0000001$name

# search for entries matching pattern :

mdb.find(db, "def", "mitochondrial", ignore.case=TRUE)

}
