% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdb.swiss.r
\docType{data}
\name{HELP.SWISS}
\alias{HELP.SWISS}
\title{lx in-memory database parser for Uniprot/Swissprot format}
\description{
read and parse Uniprot/Swissprot database in flat file format
and hold results in memory

the main functions are : \code{\link{mdb.swiss.read}} and \code{\link{mdb.swiss.load}}
}
\note{
these functions hold all the database in memory and are therefore not intented for
large databases
}
\examples{

db <- mdb.swiss.read(lx.system.file("samples/test_swiss.dat", "xlx"))

# get entry names
names(db)

# get info about specific entry :
db$P04395
db$P04395$OC
db$P04395$DR$PROSITE

# search for entries matching pattern :
mdb.find(db, "KW", "gluconate", ignore.case=TRUE)

}
