% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.dna.r
\name{length}
\alias{length}
\alias{length.Dna}
\title{Length method for Dna}
\usage{
\method{length}{Dna}(x)
}
\arguments{
\item{x}{a Dna object}
}
\value{
length of Dna sequence
}
\description{
Length of Dna sequence
}
