% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdb.base.r
\name{mdb.find}
\alias{mdb.find}
\title{grep pattern in specified records of database}
\usage{
mdb.find(db, key, pat, regex = TRUE, ignore.case = FALSE,
  full.match = FALSE)
}
\arguments{
\item{db}{a flat db or a structured db (in the later case the actual db used is db$db)}

\item{key}{record key to search in}

\item{pat}{pattern to search for}

\item{regex}{pat is a regular expression}

\item{ignore.case}{ignore case during search}

\item{full.match}{pattern should span all the key value.
equivalent to '^pattern$' but when used with regex=FALSE and ignore.case=FALSE
the engine may use precompiled entries for speedup}
}
\value{
vector (possible of 0 length) of records id (as strings)
}
\description{
grep pattern in specified records of database
}
\examples{
db <- mdb.swiss.load(lx.system.file("samples/test_swiss", "xlx"))
mdb.find(db, "KW", "gluconate", ignore.case=TRUE)
}
\seealso{
\code{\link{mdb.swiss.load}} \code{\link{mdb.obo.load}}
}
