% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdb.gaf.r
\name{mdb.gaf.filter}
\alias{mdb.gaf.filter}
\title{filter gaf table and keep only geneID <-> termID associations}
\usage{
mdb.gaf.filter(gaf, DB = c("User", "UniProtKB"), GN = NULL,
  gid.col = "DB.Object.Symbol", tid.col = "GO.ID", no.qual = TRUE)
}
\arguments{
\item{gaf}{gaf dataframe (from \link{mdb.gaf.read})}

\item{DB}{source DBs to keep (keep all if NULL)}

\item{GN}{gene IDs to keep (keep all if NULL)}

\item{gid.col}{column index of geneID}

\item{tid.col}{column index of termID}

\item{no.qual}{if TRUE (dft) remove entries with (non empty) qualifiers
else keep all entries, including the repugnant \bold{NOT} qualifier,
(this may therefore lead to plain wrong associations)}
}
\value{
a dataframe with two columns: 'gid', 'tid' specifying the
        association
}
\description{
filter gaf table and keep only geneID <-> termID associations
}
\examples{
tab <- mdb.gaf.read(lx.system.file("samples/test_gaf.dat", "xlx"))
tab <- mdb.gaf.filter(tab)
tab[tab$gid=="VPS4A",]
}
\seealso{
\link{mdb.gaf.read}
}
