% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdb.gaf.r
\name{mdb.gaf.read}
\alias{mdb.gaf.read}
\title{read Gene Association (GAF) file}
\usage{
mdb.gaf.read(pathname)
}
\arguments{
\item{pathname}{pathname of GAF file to read}
}
\value{
a dataframe with columns corresponding to
GAF 1.0 or 2.0 specifications (see details)
}
\description{
read Gene Association (GAF) file
}
\details{
GAF 1.0 or 2.0 specifications\cr
\tabular{llllll}{
\bold{Column} \tab \bold{Content} \tab \bold{Required?} \tab \bold{Cardinality}
\tab \bold{Example} \tab \bold{Dataframe_Colname}\cr
1 \tab DB \tab required \tab 1 \tab UniProtKB \tab DB\cr
2 \tab DB Object ID \tab required \tab 1 \tab P12345 \tab DB.Object.ID\cr
3 \tab DB Object Symbol \tab required \tab 1 \tab PHO3 \tab DB.Object.Symbol\cr
4 \tab Qualifier \tab optional \tab 0 or greater \tab NOT \tab Qualifier\cr
5 \tab GO ID \tab required \tab 1 \tab GO:0003993 \tab GO.ID\cr
6 \tab DB:Reference (|DB:Reference) \tab required \tab 1 or greater \tab PMID:2676709 \tab DB.Reference\cr
7 \tab Evidence Code \tab required \tab 1 \tab IMP \tab Evidence.Code\cr
8 \tab With (or) From \tab optional \tab 0 or greater \tab GO:0000346 \tab With.or.From\cr
9 \tab Aspect \tab required \tab 1 \tab F \tab Aspect\cr
10 \tab DB Object Name \tab optional \tab 0 or 1 \tab Toll-like receptor 4 \tab DB.Object.Name\cr
11 \tab DB Object Synonym (|Synonym) \tab optional \tab 0 or greater \tab hToll|Tollbooth \tab DB.Object.Synonym\cr
12 \tab DB Object Type \tab required \tab 1 \tab protein \tab DB.Object.Type\cr
13 \tab Taxon(|taxon) \tab required \tab 1 or 2 \tab taxon:9606 \tab Taxon\cr
14 \tab Date \tab required \tab 1 \tab 20090118 \tab Date\cr
15 \tab Assigned By \tab required \tab 1 \tab SGD \tab Assigned.By\cr
16 \tab Annotation Extension \tab optional \tab 0 or greater \tab part_of(CL:0000576) \tab Annotation.Extension\cr
17 \tab Gene Product Form ID \tab optional \tab 0 or 1 \tab UniProtKB:P12345-2 \tab Gene.Product.Form.ID\cr
}
}
\note{
the GAF file may be provided in plain text or gzipped format. this is checked
automaticaly, there is no need to add the .gz extension.
}
\examples{
tab <- mdb.gaf.read(lx.system.file("samples/test_gaf.dat", "xlx"))
tab <- mdb.gaf.filter(tab)
tab[tab$gid=="VPS4A",]

tac <- mdb.gaf.read(lx.system.file("samples/test_gaf_compressed.dat", "xlx"))
identical(tab, tac)
}
\seealso{
\link{mdb.gaf.filter}
}
