% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdb.obo.r
\name{mdb.obo.get.ancestors}
\alias{mdb.obo.get.ancestors}
\title{get ancestors of go.id(s)}
\usage{
mdb.obo.get.ancestors(db, go.id, max.depth = Inf)
}
\arguments{
\item{db}{go db opened by \link{mdb.obo.read} or \link{mdb.obo.load}}

\item{go.id}{entry id}

\item{max.depth}{maximum depth of ancestor}
}
\value{
list of ancestors go.id's
}
\description{
get ancestors of go.id(s)
}
\examples{

db <- mdb.obo.load(lx.system.file("samples/test_obo", "xlx"))
mdb.obo.get.ancestors(db, "GO.0000083")
}
\seealso{
\link{mdb.obo.index.ancestors}
}
