% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdb.obo.r
\name{mdb.obo.index.ancestors}
\alias{mdb.obo.index.ancestors}
\title{get indexed array of ancestors}
\usage{
mdb.obo.index.ancestors(db, goids = names(db), restrict = TRUE)
}
\arguments{
\item{db}{go db opened by mdb.obo.read or mdb.obo.load}

\item{goids}{set of goid's to index}

\item{restrict}{result to goids only (do not include entries for their ancestors)}
}
\value{
named list of ancestors for each goids (and optionally their ancestors)
}
\description{
get a list indexed by goids, giving for each entry
the list of ancestors for this goid.\cr
with goids=names(db) or restrict=TRUE, this is formally equivalent to
but much quicker than :\cr
\code{sapply(goids, function(x) mdb.obo.get.ancestors(db, x))}\cr
with restrict=FALSE the resulting list includes entries
for goids as well their ancestors ids.
}
\examples{
db <- mdb.obo.load(lx.system.file("samples/test_obo", "xlx"))
anc <- mdb.obo.index.ancestors(db)
anc["GO.0000083"]
mdb.obo.get.ancestors(db, "GO.0000083")
}
\seealso{
\link{mdb.obo.get.ancestors}
}
