% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdb.obo.r
\name{mdb.obo.load}
\alias{mdb.obo.load}
\title{quick load obo db}
\usage{
mdb.obo.load(dbname, force = FALSE, local = TRUE)
}
\arguments{
\item{dbname}{filename (without extension) of obo flatfile}

\item{force}{force read and serialize even if serialized file already exists}

\item{local}{if TRUE (default), serialized DB is saved and/or loaded in
current directory else in \code{dirname(dbname)}.}
}
\value{
a list indexed by GO terms (under the form GO.<number> not GO:<number>)\cr
        each element is a list indexed by the record key\cr
        each recordkey element is either the raw line(s) or the result of a specific parser\cr
        current parsers are provided for : id, is_a, relationship and xref
}
\description{
this is a quicker version of \link{mdb.obo.read}\cr
\code{mdb.obo.load} try to recover a previously loaded and serialized 
file called : dbname.rds\cr
if it does not exist then it reads the flat file called : 'dbname.dat'
and further serialize the result into dbname.rds\cr
you may force to ignore the serialized version by using force=TRUE
}
\note{
you may add your own function \code{.mdb.obo.parse.<key>(rec)}
to parse other keys than (id, is_a, relationship and xref).
}
\examples{

db <- mdb.obo.load(lx.system.file("samples/test_obo", "xlx"))

# get entry names
names(db)

# get info about specific entry :

db$GO.0000001
db$GO.0000001$id
db$GO.0000001$name

# search for entries matching pattern :

mdb.find(db, "def", "mitochondrial", ignore.case=TRUE)
}
\seealso{
\code{\link{mdb.obo.read}}
}
