% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdb.swiss.r
\name{mdb.swiss.parse}
\alias{mdb.swiss.parse}
\title{swiss parse driver}
\usage{
mdb.swiss.parse(key, rec)
}
\arguments{
\item{key}{key to parse (currently AC,OC,KW,DR,seq)}

\item{rec}{record to process}
}
\value{
anything that should be stored under key in record.
}
\description{
parse the content of \code{key} in record.\cr
}
\details{
this function just acts as a selector to call 
function \code{.mdb.swiss.parse.<key>} if it exists
or return record if it does not.\cr
you may add your own function \code{.mdb.swiss.parse.<key>(rec)}
to parse other keys than (AC, OC, KW, DR and seq).\cr
\code{rec} is a string containing all lines of the current record 
to be parsed (with newlines as \\n) and your function may return whatever
is appropriate (usually a list).
}
