% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.utils.r
\name{patbits}
\alias{patbits}
\title{get bitfield of pattern matches in sequence}
\usage{
patbits(seq, pat, regex = FALSE)
}
\arguments{
\item{seq}{sequence string}

\item{pat}{pattern to match}

\item{regex}{pattern \code{pat} is a regular expression (see details)}
}
\value{
bitfield (see package bit) of same size as sequence where
TRUE's indicate start positions of pattern.
}
\description{
get bitfield of pattern matches in sequence
}
\details{
if \code{regex==FALSE} then match any char in string \code{pat}
}
\examples{
seq <- "ACGTACGTAC"
x <- patbits(seq, "GC")
bit::as.which(x)
sum(x)
x <- patbits(seq, "[GC]", regex=TRUE)
x <- patbits(seq, "TAC", regex=TRUE)

}
