% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.dna.r
\name{plot}
\alias{plot}
\alias{plot.Dna}
\title{Plot method for Dna}
\usage{
\method{plot}{Dna}(x, step = ceiling(length(x)/100), compass = list(x =
  c(0L, -1L, 1L, 0L), y = c(1L, 0L, 0L, -1L)), ...)
}
\arguments{
\item{x}{Dna object}

\item{step}{walking step (in bp)}

\item{compass}{integer vector of length 4 giving the direction for (a c g t)}

\item{...}{additional arguments to \link{plot}}
}
\description{
just for fun... this is for Jean ;-)
}
