% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.dna.r
\name{print}
\alias{print}
\alias{print.Dna}
\title{Print method for Dna}
\usage{
\method{print}{Dna}(x, ...)
}
\arguments{
\item{x}{a Dna object}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print Dna sequence
}
