% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.regions.r
\name{regions.bincover}
\alias{regions.bincover}
\title{get binned coverage in regions}
\usage{
regions.bincover(handle, regions = baf2clocs(handle), binsize = 10000L,
  fun = mean, use.threads = lx.use.threads())
}
\arguments{
\item{handle}{baf file handle (as returned by \link{baf.open})}

\item{regions}{clocations regions to bin (default is regions spanning
all chromosomes)}

\item{binsize}{size of bins}

\item{fun}{function to collect cover data. see \link{baf.bin.cloc}.
default is \code{mean}.}

\item{use.threads}{(see \link{lx.use.threads})}
}
\value{
a list of length nrow(regions), each element is a numerical
vector of mean coverage in adjacent windows of size \code{binsize}
in the region.
}
\description{
loop over given \code{regions}
and collect mean coverage in adjacent windows of size \code{binsize}.
}
\examples{
baf <- baf.open(lx.system.file("samples/test.baf", "xlx"))
x <- regions.bincover(baf, binsize=1000, use.threads=FALSE)
baf.close(baf)
}
\seealso{
\link{regions.bycover.range}, \link{regions.bycover.band}
}
