% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.regions.r
\name{regions.byacgt}
\alias{regions.byacgt}
\title{get regions containing only symbols [agct]}
\usage{
regions.byacgt(handle, init = basta2clocs(handle), minreg = 10000L,
  use.threads = lx.use.threads())
}
\arguments{
\item{handle}{basta file handle (as returned by \link{basta.open})}

\item{init}{regions to start with (default is regions spanning all chromosomes)}

\item{minreg}{minimum region size}

\item{use.threads}{(see \link{lx.use.threads})}
}
\value{
a nx3 matrix of (1-based) clocations
}
\description{
loop over given \code{init} regions.
foreach of them keep regions containing only a,c,g or t's.\cr
}
\examples{
fh <- basta.open(lx.system.file("samples/test.bst", "xlx"))
x <- regions.byacgt(fh, minreg=1, use.threads=FALSE)
basta.close(fh)

}
