% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.regions.r
\name{regions.bybed}
\alias{regions.bybed}
\title{get user's defined regions from bed files}
\usage{
regions.bybed(handle, filenames, init = basta2clocs(handle),
  minreg = 1L, check = TRUE, file.stop = FALSE,
  use.threads = lx.use.threads())
}
\arguments{
\item{handle}{basta/baf file handle (as returned by \link{basta.open}
or \link{baf.open})}

\item{filenames}{a character vector of filenames}

\item{init}{regions (clocations) to start with (default is regions spanning all chromosomes)}

\item{minreg}{minimum region size (see \link{clocs.inter})}

\item{check}{check that region boundaries are correct (see \link{bed2clocs})}

\item{file.stop}{boolean, stops if a bed file is not found}

\item{use.threads}{(see \link{lx.use.threads})}
}
\value{
a nx3 matrix of (1-based) clocations
}
\description{
loop over provided bed files, intersect their regions, filter out
small regions and returns clocations.
}
\note{
the function also checks if each file exists and will skip over 
or stop on non-existing file(s)
}
\examples{
fh <- basta.open(lx.system.file("samples/test.bst", "xlx"))
bedfile <- lx.system.file("samples/test.bed", "xlx")
clocs <- regions.bybed(fh, bedfile)
# this is the same as:
clocs2 <- clocs.reduce(bed2clocs(fh, bed.read(bedfile)))
#
identical(clocs, clocs2)
basta.close(fh)

}
\seealso{
\link{bed.read}, \link{bed2clocs} and \link{basta2clocs}
}
