% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.regions.r
\name{regions.bycover.band}
\alias{regions.bycover.band}
\title{select regions in band of coverage distribution}
\usage{
regions.bycover.band(handle, init = baf2clocs(handle),
  binsize = 10000L, model = "poisson", smooth.k = c(3L, 5L, 15L, 35L,
  55L), minreg = binsize, fun = mean, keep.bins = TRUE, ...,
  use.threads = lx.use.threads())
}
\arguments{
\item{handle}{baf file handle (as returned by \link{baf.open})
(ignored if \code{bins} is provided)}

\item{init}{clocations regions to start with (default is regions spanning all
chromosomes). if \code{bins} is provided it should be the regions
used to compute bins (thru \link{regions.bincover}).}

\item{binsize}{size of bins}

\item{model}{model formula (see details).}

\item{smooth.k}{k parameter to \link{lx.smooth.median} to smooth bins before
computing distribution. (use NULL or 0 to disable smoothing)}

\item{minreg}{minimum region size (should be >= binsize)}

\item{fun}{function to collect cover. see notes (default to mean)}

\item{keep.bins}{if TRUE, \code{bins} are kept as
attributes in the result (set to FALSE to save memory)}

\item{...}{additional parameters to \code{model} user-defined function
if specified (see details)}

\item{use.threads}{(see \link{lx.use.threads})}
}
\value{
a nx3 matrix of (1-based) clocations
}
\description{
compute the distribution of (mean) coverage in all \code{init} regions,
select a band in this distribution according to different
models (see details).\cr
then loop over given \code{init} regions.
foreach of them keep regions with coverage in that band.
}
\details{
let us call \code{dist} is the distribution of coverage in all bins of
size \code{binsize}.\cr
the band of coverage [a, b] is defined by (a combination of) various models,
specified by the following formula:\cr\cr
\code{ [alpha1 '*'] model1 ['+'|'-' [alphan '*'] modeln]*}\cr\cr
where:\cr
\itemize{
\item model="median": (a,b)=median(dist)-/+alpha*mad(dist)
\item model="poisson": (a,b)=mode(dist)+/-alpha*sqrt(mode(dist))\cr
where mode(dist) is the coverage value associated to the largest maximum
of the distribution. this model correspond roughly to a poisson distributed
coverage (when coverage is large enough).
\item model="peak": a=pos-alpha*left; b=pos+alpha*right\cr
where pos, left and right are the maximum peak parameters returned
by \link{lx.peaks}.
\item model="twin": a=pos-alpha*(pos-pos.left)/2; b=pos+alpha*(pos.right-pos)/2\cr
where pos, pos.left, pos.right are respectively the maximum peak position,
the immediately left and right maxima, as returned
by \link{lx.maxima}.
\item model=<name>: a and b are defined by a user-provided function called
as .band.name(bins, alpha, ...) that should returns c(a, b)
}
in the previous formula, models are evaluated from left to right, each returning
an interval (band) that are further combined in the following way:\cr
\code{[a,b] + [c,d] = [min(a,c), max(b,d)]} (extension)\cr
\code{[a,b] - [c,d] = [max(a,c), min(b,d)]} (reduction)
}
\note{
result also contains bins (and smoothed bins) statistics
such as range, mode and density.\cr\cr
if \code{smooth.k != 0} then regions selection is performed
on smoothed bins.\cr
note that the smoothing is performed on the concatenation of all regions,
therefore ignoring chromosomes and regions boundaries. this will be
improved in the future.\cr\cr
code{fun} is an optional user function used in \link{regions.bincover}
to collect coverage.
}
\examples{
baf <- baf.open(lx.system.file("samples/test.baf", "xlx"))
x <- regions.bycover.band(baf, binsize=1000, smooth.k=3)
x <- regions.bycover.band(baf, binsize=1000, smooth.k=0, model="twin - 4 * poisson")
y <- attr(x, "binsize")
baf.close(baf)
}
\seealso{
\link{regions.bycover.range}
}
