% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.regions.r
\name{regions.bycover.range}
\alias{regions.bycover.range}
\title{select regions in range of coverage}
\usage{
regions.bycover.range(handle, init = baf2clocs(handle), bins = NULL,
  binsize = 10000L, mincover = 0, maxcover = Inf, minreg = binsize,
  fun = mean, keep.bins = TRUE, use.threads = lx.use.threads())
}
\arguments{
\item{handle}{baf file handle (as returned by \link{baf.open})
(ignored if \code{bins} is provided)}

\item{init}{clocations regions to start with (default is regions spanning all
chromosomes). if \code{bins} is provided \code{init} should be the
regions used to compute bins (thru \link{regions.bincover}).}

\item{bins}{list (of length \code{nrow(init)}) of binned coverage in regions
(as returned by \link{regions.bincover}).
if NULL this will be computed using \link{regions.bincover} with
same parameters. (see notes)}

\item{binsize}{size of bins}

\item{mincover}{minimum coverage (default 0)}

\item{maxcover}{maximum coverage (default +Inf).}

\item{minreg}{minimum region size (should be >= binsize)}

\item{fun}{function to collect cover. see notes (default to mean)}

\item{keep.bins}{if TRUE, \code{bins, binsize and binrange} are kept as
attributes in the result (set to FALSE to save memory)}

\item{use.threads}{(see \link{lx.use.threads})}
}
\value{
a nx3 matrix of (1-based) clocations
}
\description{
loop over given \code{init} regions.
foreach of them keep regions with mean coverage 
in range [mincover, maxcover].
}
\note{
the \code{bins} != NULL form is provided to avoid recomputation and
is mostly used internally (by \link{regions.bycover.band}).

code{fun} is an optional user function used in \link{regions.bincover}
to collect coverage. if provided, the range in \code{mincover, maxcover} should
scale these values. This parameter is ignored if \code{bins} are already 
provided.
}
\examples{
baf <- baf.open(lx.system.file("samples/test.baf", "xlx"))
x <- regions.bycover.range(baf, binsize=1000, mincover=1)
y <- attr(x, "binsize")
baf.close(baf)
}
\seealso{
\link{regions.bycover.band}
}
