% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.regions.r
\name{regions.bygc}
\alias{regions.bygc}
\title{get regions with specified gc content}
\usage{
regions.bygc(handle, init = basta2clocs(handle), winsize = 1000L,
  gcrange = c(0, 1), minreg = winsize,
  use.threads = lx.use.threads())
}
\arguments{
\item{handle}{basta or baf file handle (as returned by \link{basta.open}
or \link{baf.open}). see details}

\item{init}{regions to start with (default is regions spanning all chromosomes)}

\item{winsize}{window size to compute gc content}

\item{gcrange}{percent gc range (should be in [0, 1])}

\item{minreg}{minimum final region size (should be >= winsize)}

\item{use.threads}{(see \link{lx.use.threads})}
}
\value{
a nx3 matrix of (1-based) clocations
}
\description{
loop over given \code{init} regions.
foreach of them keep regions of specified %gc.\cr
}
\details{
there is a slight difference in the way the gc content is computed
depending whether you pass a basta or baf file handle.\cr
if a basta file handle is provided then the gc content is computed
on the basis of the reference genome.\cr
if a baf file handle is provided then the gc content is computed
on the basis of the actual observed alleles (see \link{baf.bin.cloc}).
note that this may lead to 0 counts in region with no mapping.\cr\cr
basta mode is (about 10 times) quicker than baf mode.
}
\examples{
fh <- basta.open(lx.system.file("samples/test.bst", "xlx"))
x <- regions.bygc(fh, winsize=3, gcrange=c(0., 0.5))
basta.close(fh)
}
\seealso{
\link{regions.strata.bygc} for a stratified version
}
