% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.regions.r
\name{regions.exclude}
\alias{regions.exclude}
\title{exclude locations from regions}
\usage{
regions.exclude(handle, coords, init = basta2clocs(handle),
  spaceleft = 0L, spaceright = spaceleft, minreg = 1L,
  use.threads = lx.use.threads())
}
\arguments{
\item{handle}{basta or baf file handle (as returned by \link{basta.open} 
or \link{baf.open}}

\item{coords}{nx2 matrix of absolute coordinates (1-based) or
vector of n absolute point coordinates to remove}

\item{init}{regions to start with (default is regions spanning all chromosomes)}

\item{spaceleft}{size to remove on left side of coords}

\item{spaceright}{size to remove on right side of coords}

\item{minreg}{minimum final region size}

\item{use.threads}{(see \link{lx.use.threads})}
}
\value{
a nx3 matrix of (1-based) clocations
}
\description{
remove locations (+/- margin) from regions
and keep only regions whose size is >= minreg
}
