% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.regions.r
\name{regions.strata.bygc}
\alias{regions.strata.bygc}
\title{stratify regions by gc content}
\usage{
regions.strata.bygc(handle, init = basta2clocs(handle),
  winsize = 1000L, nstrata = 5L, minreg = winsize,
  use.threads = lx.use.threads())
}
\arguments{
\item{handle}{basta or baf file handle (as returned by \link{basta.open}
or \link{baf.open}). see details}

\item{init}{regions to start with (default is regions spanning all chromosomes)}

\item{winsize}{window size to compute gc content}

\item{nstrata}{number of \%gc strata (strata go from 0. to 1. by 1/nstrata)}

\item{minreg}{minimum final region size (should be >= winsize)}

\item{use.threads}{(see \link{lx.use.threads})}
}
\value{
a vector of size nstrata. each element is a nx3 matrix of (1-based) clocations
}
\description{
stratify subregions from \code{init} regions into gc content
}
\details{
there is a slight difference in the way the gc content is computed
depending whether you pass a basta or baf file handle.\cr
if a basta file handle is provided then the gc content is computed
on the basis of the reference genome.\cr
if a baf file handle is provided then the gc content is computed
on the basis of the actual observed alleles (see \link{baf.bin.cloc}).
note that this may lead to 0 counts in region with no mapping.\cr
basta mode is (about 10 times) quicker than baf mode.
}
\examples{
fh <- basta.open(lx.system.file("samples/test.bst", "xlx"))
x <- regions.strata.bygc(fh, winsize=3)
basta.close(fh)
}
\seealso{
\link{regions.bygc} for a non stratified version
}
