% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.regions.r
\name{regions.trim}
\alias{regions.trim}
\title{trim regions}
\usage{
regions.trim(regions, trim = 1000L, minreg = 10000L)
}
\arguments{
\item{regions}{regions to trim}

\item{trim}{size to remove on both ends}

\item{minreg}{minimum final region size (should be >= trim)}
}
\value{
a nx3 matrix of (1-based) clocations
}
\description{
remove \code{trim} on both sides of regions
and keep only regions whose size is >= minreg
}
