% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.dna.r
\name{rev}
\alias{rev}
\alias{rev.Dna}
\title{Rev method for Dna}
\usage{
\method{rev}{Dna}(x)
}
\arguments{
\item{x}{Dna object}
}
\value{
Dna sequence reversed (but not complemented)
}
\description{
Reverse method for Dna
}
\examples{
x <- Dna("acgtnry")
rev(x)

}
\seealso{
\link{revcompl}
}
