% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.dna.r
\name{revcompl.Dna}
\alias{revcompl.Dna}
\title{Reverse Complement Dna subsequence}
\usage{
\method{revcompl}{Dna}(obj)
}
\arguments{
\item{obj}{a Dna sequence to reverse complement}
}
\description{
Reverse Complement Dna subsequence
}
\examples{
x <- Dna("acgtnry")
revcompl(x)

}
\seealso{
\link{compl.Dna}
}
