% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.dna.r
\name{revcompl}
\alias{revcompl}
\title{Generic method to reverse complement Dna subsequence}
\usage{
revcompl(obj)
}
\arguments{
\item{obj}{a Dna sequence to reverse complement}
}
\description{
Generic method to reverse complement Dna subsequence
}
\seealso{
\link{revcompl.Dna}, \link{compl}
}
