% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.coords.r
\name{runs2clocs}
\alias{runs2clocs}
\title{find runs of TRUE's in bitfield}
\usage{
runs2clocs(bit, chr = 0, minsize = 1L, p0 = 1L, delta = 1L)
}
\arguments{
\item{bit}{a bitfield (see package bit)}

\item{chr}{default chrindex}

\item{minsize}{minimum number of consecutive TRUE to report (see details)}

\item{p0}{region origin (see details)}

\item{delta}{region size factor (see details)}
}
\value{
nx3 matrix of clocations
}
\description{
considering a single bitfield (usually representing
allowed positions on a chromosome), this function
will recover all runs of TRUE (larger than the given
threshold) and return them as a nx3 matrix of clocations
(with specified chromosome index \code{chr}).
}
\details{
\code{p0} and \code{delta} are two parameters to transform indices
in bitfileds into actual positions on chromosomes according to:\cr
 \code{       pos = p0 + (i-1) * delta}\cr
this is useful when indices actually correspond to binned values
(delta=binsize) or to regions that do not start at 1
(p0 = from).\cr
when using delta!=1, the minsize parameter is interpreted with the
transformation applied (e.g with delta=1000 and minsize=1000, a single
TRUE will actually pass the test)
}
\examples{
b <- bit::as.bit(c(TRUE,FALSE,TRUE,TRUE,TRUE,FALSE,TRUE,TRUE,FALSE))
runs2clocs(b)
runs2clocs(b, minsize=3)
runs2clocs(b, delta=1000, minsize=1000) 
runs2clocs(b, delta=1000, minsize=2000) 

}
\seealso{
\link{bits2clocs} for a list version
}
