% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.coords.r
\name{sloc2cloc}
\alias{sloc2cloc}
\title{transform relative slocation to relative clocation}
\usage{
sloc2cloc(handle, slocation, zero.based.loc = FALSE)
}
\arguments{
\item{handle}{basta/baf file handle (as returned by \link{basta.open}
or \link{baf.open})}

\item{slocation}{relative slocation ("chrname:from-to")}

\item{zero.based.loc}{given slocation is 0-based}
}
\value{
relative clocation (1-based), NULL on error
}
\description{
see \link{HELP.COORD} for help on coordinates systems
}
\examples{
fh <- basta.open(lx.system.file("samples/test.bst", "xlx"))
sloc2cloc(fh, "seq1:1-10")
basta.close(fh)

}
