% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.coords.r
\name{sloc2coord}
\alias{sloc2coord}
\title{transform relative slocation to absolute coordinates}
\usage{
sloc2coord(handle, slocation, zero.based.loc = FALSE, truncate = TRUE)
}
\arguments{
\item{handle}{basta/baf file handle (as returned by \link{basta.open}
or \link{baf.open})}

\item{slocation}{relative slocation ("chrname:from-to")}

\item{zero.based.loc}{given slocation is 0-based}

\item{truncate}{truncate 3' to seq.size if needed}
}
\value{
absolute coordinates c(absfrom, absto) (1-based), NULL on error
}
\description{
see \link{HELP.COORD} for help on coordinates systems
}
\examples{
fh <- basta.open(lx.system.file("samples/test.bst", "xlx"))
sloc2coord(fh, "seq1:1-10")
sloc2coord(fh, "seq2:1-10")
basta.close(fh)

}
