% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.utils.r
\name{smooth.kalman}
\alias{smooth.kalman}
\title{smooth data using Kalman filter}
\usage{
smooth.kalman(x, f = 1, ...)
}
\arguments{
\item{x}{vector of equally spaced data (time series)}

\item{f}{parameter which controls the degree of smoothing (see details)}

\item{...}{other parameters to \link[KFAS]{KFS}}
}
\value{
named list with following fields\cr
x : the smoothed values\cr
kfs : the detailled kalman filter results (see \link[KFAS]{KFS})
}
\description{
smooth data using Kalman filter with SSMtrend model
}
\details{
the model used is \code{KFAS::SSMtrend} of degree 1 (local level) and Q (variance)
equals to \code{f} parameter. Namely \code{KFAS::SSMtrend(1, Q=list(matrix(f)))}
}
\note{
will force load of package KFAS (if available) because of a nasty bug in KFAS::SSMtrend
which prevents from using namespace.
}
\seealso{
\link[KFAS]{SSModel}, \link[KFAS]{fitSSM}, \link[KFAS]{KFS}
}
