% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.utils.r
\name{smooth.loess}
\alias{smooth.loess}
\title{smooth data using local polynomial regression}
\usage{
smooth.loess(x, y = NULL, span = 0.75, ...)
}
\arguments{
\item{x}{vector of abscissa if y != NULL or time series values if y == NULL}

\item{y}{vector of values}

\item{span}{parameter which controls the degree of smoothing (see \link{loess})}

\item{...}{other parameters to \link{loess}}
}
\value{
named list with following fields\cr
x : the original abscissa (see Details)\cr
y : the smoothed values\cr
loess : raw result from \link{loess}\cr
}
\description{
this is an alias of \link{lx.loess}
}
\details{
if just x is provided (i.e. y == NULL) then use x as values and seq_along(x)
as abscissa.
}
\examples{
x <- hist(rnorm(5000), breaks="fd", plot=FALSE)
smooth.loess(x$mids, x$counts)

}
\seealso{
\link{loess}
}
