% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.dna.r
\name{[.Dna}
\alias{[.Dna}
\title{Subscript extract method for Dna}
\usage{
\method{[}{Dna}(obj, index)
}
\arguments{
\item{obj}{Dna object}

\item{index}{any indexing expression (except for negative indices - see details)}
}
\value{
Dna subsequence
}
\description{
extract subscript from Dna sequence
}
\details{
\bold{negative indices} are not interpreted the usual way (i.e. as tail)
but as \bold{drop} (like in Python). this is more convenient to delete
symbols.
}
\examples{
x <- Dna("acgtnacgtn")
x[1:5]
x[5:1]
x[seq.int(1, 10, by=2)]
x[-3]
x[-3:-5]

}
\seealso{
\link{subseq} \link{[<-.Dna}
}
