% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.dna.r
\name{subseq.character}
\alias{subseq.character}
\title{Extract subsequence from character string}
\usage{
\method{subseq}{character}(obj, from, to)
}
\arguments{
\item{obj}{object to extract a subsequence}

\item{from}{start index (1:based, endpoint included)}

\item{to}{end index (1:based, endpoint included)}
}
\description{
this is equivalent to \link{substring}
}
