% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.dna.r
\name{subseq.Dna}
\alias{subseq.Dna}
\title{Extract Dna subsequence}
\usage{
\method{subseq}{Dna}(obj, from, to)
}
\arguments{
\item{obj}{object to extract a subsequence}

\item{from}{start index (1:based, endpoint included)}

\item{to}{end index (1:based, endpoint included)}
}
\description{
see \link{subseq}
}
\examples{
x <- Dna("acgtnacgtn")
subseq(x, 1, 3)
subseq(x, 1, -3)

}
\seealso{
\link{[.Dna}
}
