% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.dna.r
\name{subseq}
\alias{subseq}
\title{Generic method to extract subsequence}
\usage{
subseq(obj, from, to)
}
\arguments{
\item{obj}{object to extract a subsequence}

\item{from}{start index (1:based, endpoint included)}

\item{to}{end index (1:based, endpoint included)}
}
\description{
extract subsequence in range [from, to] (endpoints included)\cr
\code{to} may be omitted (in which case it equals the length of obj)\cr
\code{from} and \code{to} may be negative. they are interpreted as
\code{length - from + 1} or \code{length - to +1}.\cr
}
\note{
the term 'subsequence' is a misnommer, this is
actually a substring. So this function should be renamed
'substr' or 'substring'.\cr
see \link{[.Dna} for an actual subsequence
}
\seealso{
\link{subseq.Dna}, \link{subseq.character}
}
