% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.dna.r
\name{[<-.Dna}
\alias{[<-.Dna}
\title{Subscript replace method for Dna}
\usage{
\method{[}{Dna}(obj, index) <- value
}
\arguments{
\item{obj}{Dna object to be subscripted}

\item{index}{any indexing expression (except for negative indices - see details)}

\item{value}{a character string (recycled if necessary)}
}
\description{
replace subscript in Dna sequence
}
\details{
if \code{value} is shorter than index range, then it is recycled.\cr
if \code{value} is larger than index range, then it is truncated.\cr

\bold{negative indices} are not allowed here (since they are
interpreted as \bold{drop} see \link{[.Dna})\cr

you cannot currently use this to \bold{insert} symbol within sequence (since
value is truncated - I'll improve this in next versions). For the moment,
use \link{c} instead.
}
\examples{
x <- Dna("acgtnacgtn")
x[1:5] <- "a"
x[5:1] <- "cga"
x[seq.int(1, 10, by=2)] <- "n"

}
\seealso{
\link{[.Dna}
}
