% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.dna.r
\name{summary}
\alias{summary}
\alias{summary.Dna}
\title{Summary method for Dna}
\usage{
\method{summary}{Dna}(object, ...)
}
\arguments{
\item{object}{Dna object}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
Make a summary of Dna sequence
}
