% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlx.r
\docType{package}
\name{xlx}
\alias{xlx}
\alias{xlx-package}
\title{eXtended LX library}
\description{
Extensions to the LX base library.\cr

These utilities are currently subdivised in different subpackages:\cr
\subsection{General programming utilities}{
tbd
}

\subsection{Basta format}{
tbd
}

\subsection{Bim format}{
tbd
}
}
\details{
\tabular{ll}{
Package: \tab xlx\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-12-12\cr
License: \tab GPL\cr
}
}
\author{
Alain Viari
}
